requirements.txt
flask
flask-cors
numpy
scikit-learn
gunicorn
twilio
python-dotenv
